Replace spectacle with Reveal.js 😃
