+++ src/slides.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <title>Day One - Playground</title>
+  <title>Universe as a function</title>
   <meta charset="utf-8" />
   <meta name="apple-mobile-web-app-capable" content="yes">
   <meta name="mobile-web-app-capable" content="yes">
@@ -11,6 +11,6 @@
   <link href="https://fonts.googleapis.com/css?family=Lato:300,400,700" rel="stylesheet">
 </head>
 <body>
-  <div id="root"></div>
+  <div class="reveal" id="root"></div>
 </body>
 </html>
